+++ src/node/vardec.ts
@@ -1,24 +1,24 @@
-import { st } from "../symboltable/symbol-table";
+import { StObject, st } from "../symboltable/symbol-table";
 import { Identifier } from "./identifier";
 import { Evaluable, InterpreterNode } from "./interpreter-node";
 
 export class VarDec
-  extends InterpreterNode<null, [Identifier, Evaluable<any>]>
+  extends InterpreterNode<null, [Identifier, Evaluable<StObject>]>
   implements Evaluable<void>
 {
-  constructor(identifier: Identifier, value: Evaluable<any>) {
+  constructor(identifier: Identifier, value: Evaluable<StObject>) {
     super(null, [identifier, value]);
   }
 
   evaluate() {
     const [identifier, valueNode] = this.children;
-    const value = valueNode.evaluate();
+    const node = valueNode.evaluate();
 
     st.declare({
       name: identifier.value,
       isMutable: true,
-      type: typeof value,
-      value: value,
+      type: node.type,
+      value: node.value,
     });
   }
 }