+++ src/types/builtins.ts
@@ -1,3 +1,5 @@
 export const BuiltIns = {
   number: "number",
+  boolean: "boolean",
+  string: "string",
 };