+++ src/types/operations.ts
@@ -3,5 +3,13 @@ export const Operations = {
   minus: "minus",
   div: "div",
   multi: "multi",
+
   assign: "assign",
+
+  compare_equal: "compare_equal",
+  compare_greater: "compare_greater",
+  compare_less: "compare_lesser",
+
+  and: "and",
+  or: "or",
 };