+++ src/types/reserved-values.ts
@@ -1,0 +1,6 @@
+export const ReservedValues = {
+  boolean: {
+    verdadeiro: "verdadeiro",
+    falso: "falso",
+  },
+};