+++ src/types/reserved.ts
@@ -2,4 +2,7 @@ export const Reserved = {
   constante: "constante",
   seja: "seja",
   imprima: "imprima",
+  se: "se",
+  senao: "senao",
+  enquanto: "enquanto",
 };