+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # @cicerotcv/portuscript
 
+## 2.3.0
+
+### Minor Changes
+
+- 7aade0e: add strings
+
+## 2.2.0
+
+### Minor Changes
+
+- loop while
+
+## 2.1.0
+
+### Minor Changes
+
+- e973f83: new type: boolean; symboltable tests
+- parse conditions (se/senao)
+
 ## 2.0.0
 
 ### Major Changes